refs #874 Move Toot and Notification to organisms
