+++ src/renderer/components/organisms/Notification/Favourite.vue
