+++ src/renderer/components/organisms/Notification/Follow.vue
