--- src/renderer/store/molecules.ts
@@ -1,12 +1,0 @@
-import Toot, { TootState } from './molecules/Toot'
-
-export type MoleculesModuleState = {
-  Toot: TootState
-}
-
-export default {
-  namespaced: true,
-  modules: {
-    Toot
-  }
-}
