+++ src/renderer/store/organisms.ts
@@ -1,0 +1,12 @@
+import Toot, { TootState } from './organisms/Toot'
+
+export type OrganismsModuleState = {
+  Toot: TootState
+}
+
+export default {
+  namespaced: true,
+  modules: {
+    Toot
+  }
+}
